(maint) Separate out testing handler into reusable helper
Previously, the HTTP::Handler spec tests defined a HandlerTesting class. This
commit separates it into its own helper so it can be reused in other tests.
