+++ spec/unit/network/http/handler_spec.rb
@@ -1,5 +1,8 @@
 #! /usr/bin/env ruby
 require 'spec_helper'
+
+require 'puppet_spec/handler'
+
 require 'puppet/indirector_testing'
 
 require 'puppet/network/authorization'
@@ -17,7 +20,7 @@ describe Puppet::Network::HTTP::Handler do
     {
       :accept_header => "pson",
       :content_type_header => "text/pson",
-      :http_method => method,
+      :method => method,
       :path => path,
       :params => {},
       :client_cert => nil,
@@ -26,7 +29,7 @@ describe Puppet::Network::HTTP::Handler do
     }
   end
 
-  let(:handler) { TestingHandler.new() }
+  let(:handler) { PuppetSpec::Handler.new() }
 
   describe "the HTTP Handler" do
     def respond(text)
@@ -34,7 +37,7 @@ describe Puppet::Network::HTTP::Handler do
     end
 
     it "hands the request to the first route that matches the request path" do
-      handler = TestingHandler.new(
+      handler = PuppetSpec::Handler.new(
         Puppet::Network::HTTP::Route.path(%r{^/foo}).get(respond("skipped")),
         Puppet::Network::HTTP::Route.path(%r{^/vtest}).get(respond("used")),
         Puppet::Network::HTTP::Route.path(%r{^/vtest/foo}).get(respond("ignored")))
@@ -49,14 +52,14 @@ describe Puppet::Network::HTTP::Handler do
 
     it "raises an error if multiple routes with the same path regex are registered" do
       expect do
-        handler = TestingHandler.new(
+        handler = PuppetSpec::Handler.new(
           Puppet::Network::HTTP::Route.path(%r{^/foo}).get(respond("ignored")),
           Puppet::Network::HTTP::Route.path(%r{^/foo}).post(respond("also ignored")))
       end.to raise_error(ArgumentError)
     end
 
     it "raises an HTTP not found error if no routes match" do
-      handler = TestingHandler.new
+      handler = PuppetSpec::Handler.new
 
       req = a_request("GET", "/vtest/foo")
       res = {}
@@ -72,7 +75,7 @@ describe Puppet::Network::HTTP::Handler do
     end
 
     it "returns a structured error response with a stacktrace when the server encounters an internal error" do
-      handler = TestingHandler.new(
+      handler = PuppetSpec::Handler.new(
         Puppet::Network::HTTP::Route.path(/.*/).get(lambda { |_, _| raise StandardError.new("the sky is falling!")}))
 
       req = a_request("GET", "/vtest/foo")
@@ -106,7 +109,7 @@ describe Puppet::Network::HTTP::Handler do
       request = a_request
       request[:headers][Puppet::Network::HTTP::HEADER_ENABLE_PROFILING.downcase] = "true"
 
-      p = HandlerTestProfiler.new
+      p = PuppetSpec::HandlerProfiler.new
 
       Puppet::Util::Profiler.expects(:add_profiler).with { |profiler|
         profiler.is_a? Puppet::Util::Profiler::WallClock
@@ -171,58 +174,6 @@ describe Puppet::Network::HTTP::Handler do
       Resolv.stubs(:getname).with("1.2.3.4").raises(RuntimeError, "no such host")
 
       handler.resolve_node(:ip => "1.2.3.4").should == "1.2.3.4"
-    end
-  end
-
-  class TestingHandler
-    include Puppet::Network::HTTP::Handler
-
-    def initialize(* routes)
-      register(routes)
-    end
-
-    def set_content_type(response, format)
-     response[:content_type_header] = format
-    end
-
-    def set_response(response, body, status = 200)
-      response[:body] = body
-      response[:status] = status
-    end
-
-    def http_method(request)
-      request[:http_method]
-    end
-
-    def path(request)
-      request[:path]
-    end
-
-    def params(request)
-      request[:params]
-    end
-
-    def client_cert(request)
-      request[:client_cert]
-    end
-
-    def body(request)
-      request[:body]
-    end
-
-    def headers(request)
-      request[:headers] || {}
-    end
-  end
-
-  class HandlerTestProfiler
-    def start(metric, description)
-    end
-
-    def finish(context, metric, description)
-    end
-
-    def shutdown()
     end
   end
 end
